chore: move tapscripts to individual files
